// types/enums.ts
export enum Condition {
    NEW = "NEW",
    LIKE_NEW = "LIKE_NEW",
    GOOD = "GOOD",
    FAIR = "FAIR",
    POOR = "POOR"
}

export enum ListingCategory {
    ELECTRONICS = "ELECTRONICS",
    CLOTHING = "CLOTHING",
    BOOKS = "BOOKS",
    COURSE_MATERIALS = "COURSE_MATERIALS",
    MUSICAL_INSTRUMENTS = "MUSICAL_INSTRUMENTS",
    SPORTS_EQUIPMENT = "SPORTS_EQUIPMENT",
    TOOLS = "TOOLS",
    OTHER = "OTHER"
}

export enum ServiceDirection {
    OFFERING = "OFFERING",
    REQUESTING = "REQUESTING"
}

export enum GigCategory {
    TUTORING = "TUTORING",
    NOTES_SHARING = "NOTES_SHARING",
    ACADEMIC_WRITING = "ACADEMIC_WRITING",
    DESIGN_SERVICES = "DESIGN_SERVICES",
    CODING_HELP = "CODING_HELP",
    LANGUAGE_TRANSLATION = "LANGUAGE_TRANSLATION",
    EVENT_PLANNING = "EVENT_PLANNING",
    PHOTOGRAPHY = "PHOTOGRAPHY",
    MUSIC_LESSONS = "MUSIC_LESSONS",
    RESEARCH_ASSISTANCE = "RESEARCH_ASSISTANCE",
    EXAM_PREP = "EXAM_PREP",
    RESUME_WRITING = "RESUME_WRITING",
    CAMPUS_DELIVERY = "CAMPUS_DELIVERY",
    TECHNICAL_REPAIR = "TECHNICAL_REPAIR",
    OTHER = "OTHER"
}